requirements.txt
 Django>=4.0,<4.1
 psycopg2>=2.9.6,<2.9.9
 djangorestframework>=3.13.1,<3.14.0
 django-anymail[postmark]>=10.0,<10.1
 drf-spectacular>=0.15.1,<0.16